first pass at separating example (weather widgets) into example project and out of the main set of tests
