+++ packages/golden_toolkit/example/images/sample_cloudy.png
