+++ packages/golden_toolkit/example/images/sample_cold.png
