+++ packages/golden_toolkit/example/images/sample_rain.png
