+++ packages/golden_toolkit/example/images/sample_sunny.png
