--- packages/golden_toolkit/example/lib/example.dart
@@ -1,7 +1,0 @@
-library example;
-
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
