+++ packages/golden_toolkit/example/lib/weather_widgets.dart
