+++ packages/golden_toolkit/example/test/flutter_test_config.dart
@@ -1,0 +1,24 @@
+/// ***************************************************
+/// Copyright 2019-2020 eBay Inc.
+///
+/// Use of this source code is governed by a BSD-style
+/// license that can be found in the LICENSE file or at
+/// https://opensource.org/licenses/BSD-3-Clause
+/// ***************************************************
+///
+
+import 'dart:async';
+import 'dart:io';
+
+import 'package:golden_toolkit/golden_toolkit.dart';
+
+Future<void> main(FutureOr<void> Function() testMain) async {
+  GoldenToolkit.configure(GoldenToolkitConfiguration(
+    // Currently, goldens are not generated/validated in CI for this repo. We have settled on the goldens for this package
+    // being captured/validated by developers running on MacOSX. We may revisit this in the future if there is a reason to invest
+    // in more sophistication
+    skipGoldenAssertion: () => !Platform.isMacOS,
+  ));
+  await loadAppFonts();
+  return testMain();
+}
