+++ packages/golden_toolkit/example/test/goldens/all_sized_all_fonts.phone.png
