+++ packages/golden_toolkit/example/test/goldens/example_of_images_not_properly_loading.phone.png
