+++ packages/golden_toolkit/example/test/goldens/single_weather_card.png
