+++ packages/golden_toolkit/example/test/goldens/weather_accessibility.png
