+++ packages/golden_toolkit/example/test/goldens/weather_forecast.phone.png
