+++ packages/golden_toolkit/example/test/goldens/weather_forecast.tablet_landscape.png
