+++ packages/golden_toolkit/example/test/goldens/weather_image_async_load_correct_duration.phone.png
