+++ packages/golden_toolkit/example/test/goldens/weather_types_column.png
