+++ packages/golden_toolkit/pubspec.yaml
@@ -23,8 +23,6 @@ dev_dependencies:
 # needed so the images can be loaded in the example tests
 flutter:
   uses-material-design: true
-  assets:
-    - images/
   fonts:
     - family: Roboto
       fonts:
